\name{categoricalVariable}
\alias{categoricalVariable}

\docType{data}
\title{
A categorical Feature.
}

\usage{data("categoricalVariable")}

\description{
Character vector of length 391029 with five different labels.

}

\examples{
data(categoricalVariable)
unique(categoricalVariable)
}
\keyword{datasets}
\keyword{categorical}
\keyword{categoricalVariable}
