\name{FundamentalData_Q1_2018}
\alias{FundamentalData_Q1_2018}
\docType{data}
\title{
Fundamental Data of the 1st Quarter in 2018
}
\description{
This dataset was extracted out of Yahoo finance and was investigated in [Thrun et al., 2019] and clustered in [Thrun, 2019].
}
\usage{data("FundamentalData_Q1_2018")}
\format{
  The format is:
List of 3
 $ Data :'data.frame':	269 obs. of  45 variables:
  ..$ TotalRevenue                         : num [1:269] 3779000 78225 48220 63726 3084 ...
  ..$ CostofRevenue                        : num [1:269] 2348000 60835 26174 35203 882 ...
  ..$ GrossProfit                          : num [1:269] 1431000 17390 22046 28523 2202 ...
  ..$ SellingGeneralandAdministrative      : num [1:269] 459000 NaN 15162 17072 2005 ...
  ..$ Others                               : num [1:269] -3000 10272 -52 3131 1784 ...
  ..$ TotalOperatingExpenses               : num [1:269] 2872000 73833 41284 56787 5081 ...
  ..$ OperatingIncomeorLoss                : num [1:269] 907000 4392 6936 6939 -1997 ...
  ..$ TotalOtherIncomeDIVxpensesNet        : num [1:269] -28000 -344 1 -210 -240 ...
  ..$ EarningsBeforeInterestandTaxes       : num [1:269] 907000 4392 6936 6939 -1997 ...
  ..$ InterestExpense                      : num [1:269] -20000 -415 NaN -243 -238 ...
  ..$ IncomeBeforeTax                      : num [1:269] 879000 4048 6937 6729 -2237 ...
  ..$ IncomeTaxExpense                     : num [1:269] 233000 1365 2188 1896 7 ...
  ..$ NetIncomeFromContinuingOps           : num [1:269] 646000 2683 4749 4833 -2244 ...
  ..$ NetIncome_x                          : num [1:269] 644000 2817 4645 4833 -2244 ...
  ..$ NetIncome                            : num [1:269] 644000 2817 4645 4833 -2244 ...
  ..$ CashAndCashEquivalents               : num [1:269] 926000 29047 45911 94859 11217 ...
  ..$ NetReceivables                       : num [1:269] 2527000 46171 20774 151952 2774 ...
  ..$ Inventory                            : num [1:269] 2011000 471 NaN 10572 8924 ...
  ..$ TotalCurrentAssets                   : num [1:269] 5674000 80224 68061 267187 25989 ...
  ..$ LongTermInvestments                  : num [1:269] 234000 450 NaN 4155 872 ...
  ..$ PropertyPlantandEquipment            : num [1:269] 4216000 14561 3093 32247 7073 ...
  ..$ IntangibleAssets                     : num [1:269] 78000 40706 3975 6169 125 ...
  ..$ OtherAssets                          : num [1:269] 810000 8224 1091 2978 13310 ...
  ..$ DeferredLongTermAssetCharges         : num [1:269] 759000 684 1091 784 1405 ...
  ..$ TotalAssets                          : num [1:269] 11262000 167807 83155 351220 47369 ...
  ..$ AccountsPayable                      : num [1:269] 1442000 10567 1698 17316 1386 ...
  ..$ ShortDIVurrentLongTermDebt           : num [1:269] 1275000 30192 NaN 26668 917 ...
  ..$ OtherCurrentLiabilities              : num [1:269] 1064000 36942 22781 92297 2659 ...
  ..$ TotalCurrentLiabilities              : num [1:269] 2577000 54430 24479 114210 4299 ...
  ..$ OtherLiabilities                     : num [1:269] 1795000 19435 6876 29347 2018 ...
  ..$ TotalLiabilities                     : num [1:269] 5576000 97136 31355 165628 6980 ...
  ..$ CommonStock                          : num [1:269] 198000 14946 5198 15250 28644 ...
  ..$ RetainedEarnings                     : num [1:269] NaN 44030 34767 40374 -8965 ...
  ..$ TreasuryStock                        : num [1:269] 5455000 11686 NaN 129968 20710 ...
  ..$ OtherStockholderEquity               : num [1:269] 5455000 11686 NaN 129968 20710 ...
  ..$ TotalStockholderEquity               : num [1:269] 5653000 70662 51212 185592 40389 ...
  ..$ NetTangibleAssets                    : num [1:269] 5325000 6314 40302 140939 40264 ...
  ..$ Depreciation                         : num [1:269] 156000 2728 331 1381 410 ...
  ..$ AdjustmentsToNetIncome               : num [1:269] 216000 1911 116 2912 39 ...
  ..$ ChangesInOtherOperatingActivities    : num [1:269] -20000 -2174 -829 NaN 428 ...
  ..$ TotalCashFlowFromOperatingActivities : num [1:269] 452000 7349 4274 -8241 -1367 ...
  ..$ CapitalExpenditures                  : num [1:269] -88000 -966 -1778 -2067 -155 ...
  ..$ TotalCashFlowsFromInvestingActivities: num [1:269] 30000 -879 -1766 -2746 -484 ...
  ..$ TotalCashFlowsFromFinancingActivities: num [1:269] -789000 -6660 -21867 -961 -204 ...
  ..$ ChangeInCashandCashEquivalents       : num [1:269] -306000 -215 2508 -11842 -2062 ...
 $ Names: chr [1:269, 1:6] "1COV" "A1OS" "AAD" "AAG" ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:6] "Key" "ISIN" "Company" "Sector" ...
 $ Cls  : num [1:269] 1 1 1 1 2 1 1 1 3 1 ...
}
\details{
Stocks are selected by the German Prime standard accoridingly to the "Names" data frame.
Fundamental Data with missing values is stored in "Data". The rownames of "Data" have the same Key as the first row of "Names" which is the trading symbol.
Cls provides the clustering as a numerical vector of 1:k classes performed by Databionic Swarm in [Thrun, 2019].

}
\source{
Yahoo finance
}
\references{
Thrun, M. C., : Knowledge Discovery in Quarterly Financial Data of Stocks Based on the Prime Standard using a Hybrid of a Swarm with SOM, in Verleysen, M. (Ed.), European Symposium on Artificial Neural Networks, Computational Intelligence and Machine Learning (ESANN), Vol. 27, pp. 397-402, Ciaco, ISBN: 978-287-587-065-0, Bruges, Belgium, 2019. 

[Thrun et al., 2019]	Thrun, M. C., Gehlert, Tino, & Ultsch, A. : Analyzing the Fine Structure of Distributions, arXiv:1908.06081, 2019. 
}
\examples{
data(FundamentalData_Q1_2018)
## maybe str(FundamentalData_Q1_2018) ; plot(FundamentalData_Q1_2018) ...
}
\keyword{datasets}
