\name{GermanPostalCodesShapes}
\alias{GermanPostalCodesShapes}
\docType{data}
\title{
GermanPostalCodesShapes
}
\description{
GermanPostalCodesShapes
}
\usage{data("GermanPostalCodesShapes")}

\details{
GermanPostalCodesShapes
}
\source{
You could read \url{https://www.r-bloggers.com/case-study-mapping-german-zip-codes-in-r/}, if you want to change the map.
}

\examples{
data(GermanPostalCodesShapes)
str(GermanPostalCodesShapes)
}
\keyword{datasets}
