\name{HeatmapColors}
\alias{HeatmapColors}
\docType{data}
\title{
Default color sequence for plots
}
\description{
Defines the default color sequence for plots made with PixelMatrixPlot
}
\usage{data("HeatmapColors")}
\format{
  A vector with different strings describing colors for this plot.
}