\name{MTY}
\alias{MTY}

\docType{data}
\title{
Muncipal Income Tax Yield
}

\usage{data("MTY")}

\description{
Numerical vector of length 11194. details in [Ultsch/Behnisch, 2017; Thrun/Ultsch, 2018].

}
\references{
[Thrun/Ultsch, 2018] Thrun, M. C., & Ultsch, A. : Effects of the payout system of income taxes to municipalities in Germany, in Papiez, M. & Smiech,, S. (eds.), Proc. 12th Professor Aleksander Zelias International Conference on Modelling and Forecasting of Socio-Economic Phenomena, pp. 533-542, Cracow: Foundation of the Cracow University of Economics, Cracow, Poland, 2018. 

[Ultsch/Behnisch, 2017]	 Ultsch, A., Behnisch, M.: Effects of the payout system of income taxes to municipalities in Germany, Applied Geography, Vol. 81, pp. 21-31, 2017.
}
\examples{
data(MTY)
str(MTY)
}
\keyword{datasets}
\keyword{MTY}
\keyword{Muncipal Income Tax Yield}
