\name{QQplotForStandardiziation}
\alias{QQplotForStandardiziation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
QQplot with Fit
}
\description{
QQplot with linear fit
}
\usage{
QQplotForStandardiziation(Data, TransData, xug = -3, xog = 3, xlab = "Normal", yDataLab =

"Data", yTransDataLab = "Trasformated Data", Symbol4Gerade = "red", main = "", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{
...}
  \item{TransData}{
...}
  \item{xug}{
...}
  \item{xog}{
...}
  \item{xlab}{
...}
  \item{yDataLab}{
...}
  \item{yTransDataLab}{
...}
  \item{Symbol4Gerade}{
...}
  \item{main}{
...}
  \item{\dots}{
...
}
}
\details{
...
}
\value{
plot
}
\references{
Michael, J. R.: The stabilized probability plot, Biometrika, Vol. 70(1), pp. 11-17, 1983.
}
\author{
Michael Thrun
}
\keyword{qqplot}% use one of  RShowDoc("KEYWORDS")

