\name{QQplotWithFit}
\alias{QQplotWithFit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
QQplot with Fit
}
\description{
QQplot with linear fit
}
\usage{
QQplotWithFit(Data, TransData, xug = -3, xog = 3, xlab = "Normal", yDataLab =

"Data", yTransDataLab = "Trasformated Data", Symbol4Gerade = "red", main = "", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{
...}
  \item{TransData}{
...}
  \item{xug}{
...}
  \item{xog}{
...}
  \item{xlab}{
...}
  \item{yDataLab}{
...}
  \item{yTransDataLab}{
...}
  \item{Symbol4Gerade}{
...}
  \item{main}{
...}
  \item{\dots}{
...
}
}
\details{
...
}
\value{
plot
}
\references{
michael 1986
}
\author{
unknown
}
\keyword{qqplot}% use one of  RShowDoc("KEYWORDS")

